README.md
Our unique approach to social enterprise and our global reputation for producing graduates with exceptional employability makes us your first choice for higher education.
Pal College of Technology & Management (PCTM) is situated in the foothills of Kumaon, in Haldwani. Since the inception of the state of Uttarakhand, Haldwani has been the center of attraction of not only tourists but also major corporations due to conducive economic, social and cultural conditions. Pal College of Technology & Management aims to further the educational development of the area and its outskirts under the auspices of the renowned Pal Group, as quality education is the region’s most pressing requirement.